Task/requirements.txt
Flask
celery
boto3
python-dotenv